Change how Radeon Profile Daemon's versioning
Prior versioning system was getting a ridiculous value as version number.
The new versioning is based on the Major.Minor.Patch.Build system.
- Major number shall be incremented when significant changes are made or
when a part of the application tandem is change in a way that it wouldn't
be compatible with the previous version.
- Minor number shall be incremented for new code and functionalities that
down break previous functionality, but are more than bug or issues. This
could be used when adding support for a new GPU/architecture, when
changing significantly the UI, adding new capacities to the daemon, etc.
- Patch number shall be incremented when a new release is provided that
fixes code, but doesn't introduce new functionality.
- Build number shall be represented by the last commit with a "-dirty"
note if it was modified.

The versions between Radeon Profile Daemon and Radeon Profile must be kept
compatible and coherent.

When needed, the version number will be found by including the new
"info.h" header file.

Let's change this here and in Radeon Profile. This mark the beginning of
the 2.0.0.0.

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>